+++ src/main/java/view/TemplateManager.java
@@ -1,0 +1,7 @@
+package view;
+
+import java.util.Map;
+
+public interface TemplateManager {
+    byte[] applyCompile(String viewName, Map<String, Object> modelMap);
+}
