+++ src/main/java/view/ViewResolver.java
@@ -1,16 +1,16 @@
 package view;
 
+import fileloader.TemplateFileLoader;
+
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.function.Predicate;
 
 public class ViewResolver {
-    private static final String TEMPLATES_PREFIX = "/templates";
-    private static final String TEMPLATES_SUFFIX = ".html";
-    private static final String PATH_FORMAT = "%s%s%s";
     private static final String REDIRECT = "redirect: ";
     private static final Map<Predicate<String>, Function<String, View>> VIEWS = new HashMap<>();
+    private static final TemplateManager TEMPLATE_MANAGER = new HandlebarsManager(TemplateFileLoader.getInstance());
 
     static {
         VIEWS.put(viewName -> viewName.startsWith(REDIRECT), RedirectView::new);
@@ -32,10 +32,6 @@ public class ViewResolver {
                 .filter(entry -> entry.getKey().test(viewName))
                 .map(entry -> entry.getValue().apply(viewName))
                 .findAny()
-                .orElse(new HandlebarsView(buildPath(viewName)));
-    }
-
-    private String buildPath(String viewName) {
-        return String.format(PATH_FORMAT, TEMPLATES_PREFIX, viewName, TEMPLATES_SUFFIX);
+                .orElse(new TemplateView(viewName, TEMPLATE_MANAGER));
     }
 }
